change price field

+++ main/models.py
@@ -42,7 +42,7 @@ class CustomUser(AbstractUser):
     inn = models.CharField(max_length=12, blank=True, verbose_name='ИНН')
     time_update = models.DateTimeField(auto_now=True, verbose_name='Дата изменения')
     request_limit = models.IntegerField(null=False, default=100, verbose_name='Лимит запросов')
-    cost_per_hour = models.FloatField(null=False, default=1000.00, verbose_name='Стоимость часа')
+    cost_per_hour = models.DecimalField(max_digits=15, decimal_places=2, null=False, default=1000, verbose_name='Стоимость часа')
 
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS = []
